prov/rxd: switch order of headers (and increment protocol version)
The presence of and value of num_segs in the sar_hdr are dependent on whether we have a data_hdr and tag_hdr.

Because of this, it makes more sense to include the data header and tag headers before that sar header in the packet. This allows for easier to follow code and fewer checks in the packet initialization.

Signed-off-by: aingerson <alexia.ingerson@intel.com>